requirements.txt
lxml
bs4
requests
pycryptodomex
keyring